refactor: rename job_entity and job_status
