+++ crates/core/src/entities/worker_job.rs
@@ -1,0 +1,32 @@
+use std::fmt;
+
+use serde::{Deserialize, Serialize};
+use uuid::Uuid;
+/// Identifier for a question.
+#[derive(Debug, Eq, Hash, PartialEq, Serialize, Deserialize, Clone)]
+pub struct JobId(pub Uuid);
+
+#[derive(Serialize, Deserialize, PartialEq, Debug)]
+pub struct WorkerJob {
+    pub id: JobId,
+    pub customer_id: String,
+    pub cairo_job_key: String,
+    pub offchain_proof: bool,
+    pub proof_layout: String,
+    pub cairo_pie: String,
+}
+
+#[derive(Serialize, Deserialize, PartialEq, Debug)]
+pub struct NewWorkerJob {
+    pub customer_id: String,
+    pub cairo_job_key: String,
+    pub offchain_proof: bool,
+    pub proof_layout: String,
+    pub cairo_pie: String,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
+pub struct JobResponse {
+    pub code: Option<String>,
+    pub message: Option<String>,
+}
