+++ crates/core/src/ports/job.rs
@@ -1,12 +1,12 @@
 use async_trait::async_trait;
 
 use crate::common::core_error::CoreError;
-use crate::entities::job::{JobEntity, JobId};
+use crate::entities::job::{StatusEntity, StatusId};
 
 #[async_trait]
-pub trait JobPort {
-    async fn add(&self, job: JobEntity) -> Result<JobEntity, CoreError>;
-    async fn update(&self, job: JobEntity) -> Result<JobEntity, CoreError>;
-    async fn delete(&self, job_id: &JobId) -> Result<(), CoreError>;
-    async fn get(&self, job_id: &JobId) -> Result<JobEntity, CoreError>;
+pub trait StatusPort {
+    async fn add(&self, job: StatusEntity) -> Result<StatusEntity, CoreError>;
+    async fn update(&self, job: StatusEntity) -> Result<StatusEntity, CoreError>;
+    async fn delete(&self, job_id: &StatusId) -> Result<(), CoreError>;
+    async fn get(&self, job_id: &StatusId) -> Result<StatusEntity, CoreError>;
 }
